Separate filter pipeline functions to filter_pipeline.py module
